requirements.txt
apache-airflow==2.2.3
pytest
